REFACTOR: move binning utils to spec.
